+++ src/BoundaryConditions/BoundaryCondition.cpp
@@ -72,7 +72,7 @@ void BoundaryCondition::deleteGhostParticlesInBoundaryCells() {
             std::cout << "Had to delete ghost particle in boundary cell";
             std::vector<Particle> particles = std::vector<Particle>(
                     cell->getParticles().size() - numberOfGhostParticles);
-            for (int i = 0; i < cell->getParticles().size(); i++) {
+            for (int i = 0; i < static_cast<int>(cell->getParticles().size()); i++) {
                 if (!particles[i].isGhostParticle) {
                     particles[i] = cell->getParticles()[i];
                 }
